app/WEBPAGE/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
python-socketio==5.8.0
Pillow==10.0.0
opencv-python==4.8.0.76
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pytesseract==0.3.10
catboost==1.2.2
pydicom==2.4.4
requests==2.31.0
# stripe==7.6.0  # REMOVED: Payment processing disabled
inference-sdk